requirements.txt
et_xmlfile==2.0.0
openpyxl==3.1.5
pyaes==1.6.1
pyasn1==0.6.2
python-dotenv==1.2.1
rsa==4.9.1
Telethon==1.42.0
